feat(report): add progress logs and significant-prefix discovery; extract format
