+++ src/helper.rs
@@ -24,6 +24,71 @@ pub fn sanitize_url(url_str: &str) -> String {
     url.to_string()
 }
 
+/// Format numbers into human-readable format
+/// Uses units K, M, B, T where B represents billion (10^9)
+/// Handles special f64 values: NaN, infinity, and negative numbers
+pub fn format_number(num: f64) -> String {
+    // Handle special f64 values
+    if num.is_nan() {
+        return "NaN".to_string();
+    }
+    if num.is_infinite() {
+        return if num.is_sign_positive() {
+            "inf".to_string()
+        } else {
+            "-inf".to_string()
+        };
+    }
+    if num == 0.0 {
+        return "0".to_string();
+    }
+
+    // Handle negative numbers by formatting absolute value and adding sign back
+    let is_negative = num < 0.0;
+    let abs_num = num.abs();
+
+    let mut suffix = "";
+    let mut divisor = 1.0;
+
+    // Define units, using B as billion (10^9)
+    let units = [
+        (1_000_000_000_000.0, "T"), // trillion
+        (1_000_000_000.0, "B"),     // billion (not G)
+        (1_000_000.0, "M"),         // million
+        (1_000.0, "K"),             // thousand
+    ];
+
+    for (threshold, unit) in &units {
+        if abs_num >= *threshold {
+            suffix = unit;
+            divisor = *threshold;
+            break;
+        }
+    }
+
+    if divisor == 1.0 {
+        let result = format!("{}", abs_num as u64);
+        return if is_negative {
+            format!("-{}", result)
+        } else {
+            result
+        };
+    }
+
+    // Use integer arithmetic to avoid floating point precision issues
+    let scaled = (abs_num * 100.0) / divisor;
+    let integer_part = (scaled / 100.0) as u64;
+    let decimal_part = (scaled % 100.0) as u32;
+
+    let formatted_str = format!("{}.{:02}{}", integer_part, decimal_part, suffix);
+
+    if is_negative {
+        format!("-{}", formatted_str)
+    } else {
+        formatted_str
+    }
+}
+
 /// Calculate Codis slot for a given key
 pub fn codis_slot(key: &[u8]) -> u16 {
     let key = key
@@ -82,6 +147,119 @@ macro_rules! impl_serde_str_conversion {
 #[cfg(test)]
 mod tests {
     use super::*;
+
+    #[test]
+    fn test_format_number_special_values() {
+        // Test NaN
+        assert_eq!(format_number(f64::NAN), "NaN");
+
+        // Test positive infinity
+        assert_eq!(format_number(f64::INFINITY), "inf");
+
+        // Test negative infinity
+        assert_eq!(format_number(f64::NEG_INFINITY), "-inf");
+
+        // Test zero
+        assert_eq!(format_number(0.0), "0");
+        assert_eq!(format_number(-0.0), "0");
+    }
+
+    #[test]
+    fn test_format_number_small_numbers() {
+        // Test numbers less than 1000
+        assert_eq!(format_number(0.0), "0");
+        assert_eq!(format_number(1.0), "1");
+        assert_eq!(format_number(999.0), "999");
+
+        // Test negative small numbers
+        assert_eq!(format_number(-1.0), "-1");
+        assert_eq!(format_number(-999.0), "-999");
+    }
+
+    #[test]
+    fn test_format_number_thousand_unit() {
+        // Test K unit (1000 - 999,999)
+        assert_eq!(format_number(1000.0), "1.00K");
+        assert_eq!(format_number(1500.0), "1.50K");
+        assert_eq!(format_number(999999.0), "999.99K");
+
+        // Test negative K unit
+        assert_eq!(format_number(-1000.0), "-1.00K");
+        assert_eq!(format_number(-1500.0), "-1.50K");
+
+        // Test edge cases
+        assert_eq!(format_number(999.0), "999"); // Should not be formatted as K
+        assert_eq!(format_number(1001.0), "1.00K"); // Should be formatted as K with 2 decimals
+    }
+
+    #[test]
+    fn test_format_number_million_unit() {
+        // Test M unit (1,000,000 - 999,999,999)
+        assert_eq!(format_number(1_000_000.0), "1.00M");
+        assert_eq!(format_number(1_500_000.0), "1.50M");
+        assert_eq!(format_number(999_999_999.0), "999.99M");
+
+        // Test negative M unit
+        assert_eq!(format_number(-1_000_000.0), "-1.00M");
+        assert_eq!(format_number(-1_500_000.0), "-1.50M");
+    }
+
+    #[test]
+    fn test_format_number_billion_unit() {
+        // Test B unit (1,000,000,000 - 999,999,999,999)
+        assert_eq!(format_number(1_000_000_000.0), "1.00B");
+        assert_eq!(format_number(1_500_000_000.0), "1.50B");
+        assert_eq!(format_number(999_999_999_999.0), "999.99B");
+
+        // Test negative B unit
+        assert_eq!(format_number(-1_000_000_000.0), "-1.00B");
+        assert_eq!(format_number(-1_500_000_000.0), "-1.50B");
+    }
+
+    #[test]
+    fn test_format_number_trillion_unit() {
+        // Test T unit (>= 1,000,000,000,000)
+        assert_eq!(format_number(1_000_000_000_000.0), "1.00T");
+        assert_eq!(format_number(1_500_000_000_000.0), "1.50T");
+        assert_eq!(format_number(999_999_999_999_999.0), "999.99T");
+
+        // Test negative T unit
+        assert_eq!(format_number(-1_000_000_000_000.0), "-1.00T");
+        assert_eq!(format_number(-1_500_000_000_000.0), "-1.50T");
+    }
+
+    #[test]
+    fn test_format_number_decimal_places() {
+        // With the new logic, all formatted numbers now show 2 decimal places
+        assert_eq!(format_number(100_000.0), "100.00K");
+        assert_eq!(format_number(1_500_000.0), "1.50M");
+        assert_eq!(format_number(250_000_000.0), "250.00M");
+
+        assert_eq!(format_number(10_000.0), "10.00K");
+        assert_eq!(format_number(15_000.0), "15.00K");
+        assert_eq!(format_number(99_999.0), "99.99K");
+
+        assert_eq!(format_number(1_000.0), "1.00K");
+        assert_eq!(format_number(1_500.0), "1.50K");
+        assert_eq!(format_number(9_999.0), "9.99K");
+    }
+
+    #[test]
+    fn test_format_number_edge_cases() {
+        // Test very large numbers
+        assert_eq!(format_number(1_000_000_000_000_000.0), "1000.00T");
+
+        // Test numbers just at unit boundaries
+        assert_eq!(format_number(999_999.0), "999.99K");
+        assert_eq!(format_number(1_000_000.0), "1.00M");
+        assert_eq!(format_number(999_999_999.0), "999.99M");
+        assert_eq!(format_number(1_000_000_000.0), "1.00B");
+        assert_eq!(format_number(999_999_999_999.0), "999.99B");
+        assert_eq!(format_number(1_000_000_000_000.0), "1.00T");
+
+        // Test floating point precision
+        assert_eq!(format_number(1_000_000_000.5), "1.00B");
+    }
 
     #[test]
     fn test_sanitize_url() {
